+++ contracts/test/oracles/proto/uniswap/UniswapV2OracleStub.sol
@@ -21,10 +21,10 @@ contract UniswapV2OracleStub is UniswapV2Oracle {
 
     function stubSetObservation(
         address token,
-        uint256 price,
-        uint256 tokenLiquidity,
-        uint256 quoteTokenLiquidity,
-        uint256 timestamp
+        uint112 price,
+        uint112 tokenLiquidity,
+        uint112 quoteTokenLiquidity,
+        uint32 timestamp
     ) public {
         ObservationLibrary.Observation storage observation = observations[token];
